Drop nullToEmpty and requireNonNullOrElse
Use the nonnull getter variants where appropriate.

+++ elanmanager/impl/src/main/java/org/opendaylight/netvirt/elan/internal/ElanDpnInterfacesListener.java
@@ -9,7 +9,6 @@
 package org.opendaylight.netvirt.elan.internal;
 
 import static java.util.Collections.emptyList;
-import static org.opendaylight.netvirt.elan.utils.ElanUtils.requireNonNullElse;
 
 import java.math.BigInteger;
 import java.util.List;
@@ -78,9 +77,10 @@ public class ElanDpnInterfacesListener
         ElanInstance elanInstance = elanInstanceCache.get(elanInstanceName).orNull();
 
         if (elanInstance != null && !elanInstance.isExternal() && ElanUtils.isVlan(elanInstance)) {
-            List<String> interfaces = requireNonNullElse(update.getInterfaces(), emptyList());
+            List<String> interfaces = update.getInterfaces();
             // trigger deletion for vlan provider intf on the DPN for the vlan provider network
-            if (interfaces.size() == 1 && interfaceManager.isExternalInterface(interfaces.get(0))) {
+            if (interfaces != null && interfaces.size() == 1 && interfaceManager.isExternalInterface(
+                    interfaces.get(0))) {
                 LOG.debug("deleting vlan prv intf for elan {}, dpn {}", elanInstanceName, dpnId);
                 jobCoordinator.enqueueJob(dpnId.toString(), () -> {
                     elanService.deleteExternalElanNetwork(elanInstance, dpnId);
